emby_webhooks/requirements.txt
Flask
requests
python-dateutil
python-dotenv
Werkzeug